Overview layout complete. Components refactored into separate files.
